Added .gitignore & edits
Added .gitignore in order to ignore jupyter notebook checkpoints. Also continued revisit to jupyter notebook. Started adding occasional commentary and docstrings for functions that were previously unknown for future me, in case I fail to use them in the immediate future and revisit the notebook in the more distant future.